ai_stylist/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
gradio>=3.35.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=9.0.0
numpy>=1.24.0